requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.3.0
PyPDF2==3.0.1
pdfplumber==0.10.3
pandas==2.1.3
numpy==1.26.2
werkzeug==3.0.1

